Upgrade app with login and signup blueprint

+++ app/Blueprints/main.py
@@ -1,9 +1,10 @@
-from flask import render_template, request, Blueprint, jsonify, redirect, url_for
+from flask import render_template, request, Blueprint, jsonify, redirect, url_for, flash
 from datetime import datetime
 from flask_wtf import FlaskForm
+from flask_login import login_required, current_user
 from wtforms import (StringField, SubmitField, BooleanField, DateField) 
 from wtforms.validators import DataRequired, Length
-from app.models.todo import ToDo
+from app.models.todo import ToDo, User
 from app.extensions import db
 
 
@@ -16,6 +17,7 @@ class NewTask(FlaskForm):
 
 # Main route
 @bp.route('/main', methods=['GET', 'POST'])
+@login_required
 def to_do():
     task = None
     form = NewTask()
@@ -25,36 +27,45 @@ def to_do():
         task = form.task.data
         deadline = form.deadline.data
         
-        todo = ToDo(username='Phong', task=task, deadline=deadline, completed=False)
+        todo = ToDo(user_id=current_user.id, task=task, deadline=deadline, completed=False)
         db.session.add(todo)
         db.session.commit()
+        flash('Create task successfully', 'success')
         return redirect(url_for('main.to_do'))
     # Display to do list
-    to_do_list = ToDo.query.order_by(ToDo.id).all()
+    to_do_list = ToDo.query.filter_by(user_id=current_user.id).order_by(ToDo.id).all()
     deadline_passed = {}
     for todo in to_do_list:
         deadline_passed[todo.id] = True if datetime.now().date() > todo.deadline else False
 
 
-    # Change completed status
-    if request.method == 'POST':
-        todo_id = request.form['completed']
+    if request.method == 'POST' and request.form['button']:
+        todo_id = request.form['id']
         todo = ToDo.query.filter_by(id=todo_id).first()
-        todo.completed = not todo.completed
-        db.session.commit()
+        # Change completed status
+        if request.form['button'] == 'completed':
+            todo.completed = not todo.completed
+            db.session.commit()
+            flash('Complete status change!', 'info')
+        elif request.form['button'] == 'delete':
+        # Delete request
+            db.session.delete(todo)
+            db.session.commit()
+            flash('Deleted task', 'danger')
+            
         return redirect(url_for('main.to_do'))
-
-    # Delete request
-    if request.method == 'POST':
-        todo_id = request.form["delete"]
-        return redirect(url_for('main.del_task', todo_id=todo_id))
+    
     
     return render_template('main.html', task=task, form=form, to_do_list=to_do_list, deadline_passed=deadline_passed)
 
-# Delete a task
-@bp.route('/del/<todo_id>', methods=['GET', 'POST'])
-def del_task(todo_id):
-    todo = ToDo.query.filter_by(id=todo_id).first()
-    db.session.delete(todo)
-    db.session.commit()
-    return redirect(url_for('main.to_do'))
+# # Delete a task
+# @bp.route('/del/<todo_id>', methods=['GET', 'POST'])
+# @login_required
+# def del_task(todo_id):
+#     if request.method == 'POST':
+#         todo = ToDo.query.filter_by(id=todo_id).first()
+#         db.session.delete(todo)
+#         db.session.commit()
+#         return redirect(url_for('main.to_do'))
+#     else:
+#         return render_template('denied.html')
